add pundit policies to authorized post actions

+++ app/controllers/posts_controller.rb
@@ -1,7 +1,7 @@
 class PostsController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
   before_action :set_post, only: [:show, :edit, :update, :destroy]
-  before_action :validate_post_owner, only: [:edit, :update, :destroy]
+  # before_action :validate_post_owner, only: [:edit, :update, :destroy]
 
   def index
     @posts = Post.includes(:categories, :user).all
@@ -27,9 +27,12 @@ class PostsController < ApplicationController
     @post = Post.find_by(short_url: params[:short_url])
   end
 
-  def edit; end
+  def edit
+    authorize @post, :edit?, policy_class: PostPolicy
+  end
 
   def update
+    authorize @post, :update?, policy_class: PostPolicy
     if @post.update(post_params)
       redirect_to posts_path
     else
@@ -38,6 +41,7 @@ class PostsController < ApplicationController
   end
 
   def destroy
+    authorize @post, :destroy?, policy_class: PostPolicy
     unless @post.destroy
       flash[:alert] = 'Cannot be delete, because this post has comments.'
     end
@@ -45,13 +49,12 @@ class PostsController < ApplicationController
   end
 
   private
-
-  def validate_post_owner
-    unless @post.user == current_user
-      flash[:notice] = 'the post not belongs to you'
-      redirect_to posts_path
-    end
-  end
+  # def validate_post_owner
+  #   unless @post.user == current_user
+  #     flash[:notice] = 'the post not belongs to you'
+  #     redirect_to posts_path
+  #   end
+  # end
 
   # def set_post
   #   @post = Post.find(params[:id])
